+++ quests.py
@@ -4,7 +4,7 @@
 class Quest():
     "quest class"
     def __init__(self, name, description, completionStatus, 
-                 requirement, reward):
+                 requirement, reward, type):
         '''
         sets up the quest class
         name: the name of the quest
@@ -12,12 +12,14 @@ class Quest():
         completionStatus: a bool of whether or not the quest is complete
         requirement: the number of required resources
         reward: the number of coins you get after you complete the quest
+        type: the type of quest
         '''
         self.name = name
         self.description = description
         self.completionStatus = completionStatus
         self.requirement = requirement
         self.reward = reward
+        self.type = type
 
     def printDescription(self):
         '''prints the description of the quest'''
@@ -30,3 +32,5 @@ class Quest():
             questStatus = (
                 f"\033[38;2;{224};{13};{13}m{'incomplete'}\033[0m")
         print(f"this quest is {questStatus}")
+
+